refactor: moved recipe-related components to a single module
