+++ src/front/modules/recipes/recipe-subcomponents/AddToCollectionModal.jsx
